swaybar/dbusmenu: don't truncate pointer position to int

+++ include/swaybar/tray/dbusmenu.h
@@ -6,7 +6,7 @@
 
 void swaybar_dbusmenu_open(struct swaybar_sni *sni,
 		struct swaybar_output *output, struct swaybar_seat *seat, uint32_t serial,
-		int x, int y);
+		double x, double y);
 
 bool dbusmenu_pointer_button(void *data, struct wl_pointer *wl_pointer,
 		uint32_t serial, uint32_t time_, uint32_t button, uint32_t state);
